app/requirements.txt
sympy
Pillow
numpy
scipy
evaluation-function-utils